+++ routes/index.js
@@ -1,0 +1,2 @@
+const teams = require("./teams")
+module.exports = {teams}
